==> notebooks/40_Data_preparation.ipynb <==
### Extract donor summaries

This step is to extract summary datasets about the donors:
1. yearly_actual_donors - has the actual donations from each county by year
2. yearly_expected_donors - This is projecting the mean donations across the entire country on the indian population in that county/year to come up with yearly projections for the donations from those counties in those years.

import pandas as pd
import numpy as np

donations = pd.read_pickle('out/21/donations.pkl')
us_states = pd.read_pickle('out/11/states_gps.pkl')
us_counties = pd.read_pickle('out/11/counties_gps.pkl')
population = pd.read_pickle('out/11/indian_population.pkl')

yearly_india_pop = population.set_index(['county_id', 'county_norm', 'state']).apply(sum)
yearly_donor_pop = donations.groupby('activity_year').donor_id.nunique()
yearly_mean_donor_fraction = yearly_donor_pop*1.0/yearly_india_pop

yearly_expected_donors = population.set_index(['county_id', 'county_norm', 'state']).mul(yearly_mean_donor_fraction, axis=1).astype('int')
yearly_expected_donors = yearly_expected_donors.reset_index()

yearly_actual_donors = donations\
    .groupby(['state', 'county_norm', 'county_id', 'activity_year'])\
    .donor_id\
    .nunique()\
    .to_frame()\
    .unstack()\
    .fillna(0)

# drop the "donor_id" dummy level
yearly_actual_donors.columns = yearly_actual_donors.columns.droplevel()

# rows don't match
yearly_expected_donors.shape, yearly_actual_donors.shape, population.shape

# To make things easier later, we want to set the values for all the counties in the yearly actual donor dataframe
yearly_actual_donors= yearly_actual_donors.reset_index()\
    .merge(population.reset_index()[['state', 'county_norm', 'county_id']],
           on=['state', 'county_norm','county_id'],
           how='right')\
    .sort_values(by='county_norm')\
    .fillna(0)


# rows match after conversion
yearly_expected_donors.shape, yearly_actual_donors.shape, population.shape

us_counties.head()

cols = range(2001, 2017)
cols.extend(['state', 'county_norm', 'county_id'])

yearly_expected_donors = yearly_expected_donors\
    .sort_values(by=['county_id'])\
    .reset_index(drop=True)[cols]

yearly_actual_donors = yearly_actual_donors\
    .sort_values(by=['county_id'])\
    .reset_index(drop=True)[cols]

yearly_actual_donors.head()

!mkdir -p out/40
yearly_expected_donors.to_pickle('out/40/yearly_expected_donors.pkl')
yearly_actual_donors.to_pickle('out/40/yearly_actual_donors.pkl')

### Testing out queries for the web app

indian_population = population

yearly_actual_donors.head()

statesfilter = ['CA', 'WA']
yearly_actual_donors[yearly_actual_donors.state.isin(statesfilter)]
yearly_expected_donors[yearly_expected_donors.state.isin(statesfilter)]
_ = indian_population[indian_population.state.isin(statesfilter)]

yearly_expected_donors.head()

actual =  yearly_actual_donors[yearly_actual_donors.state.isin(statesfilter)]
expected =  yearly_expected_donors[yearly_expected_donors.state.isin(statesfilter)]
population =  indian_population[indian_population.state.isin(statesfilter)]

difference = actual.set_index(['state', 'county_norm']) - expected.set_index(['state', 'county_norm'])

population[population.county_norm=='snohomish']

population =  indian_population[indian_population.state.isin(statesfilter)]

year = 2014
statesfilter = ['AZ', 'CA', 'CO', 'ID', 'MT', 'NM', 'NV', 'OR', 'UT', 'WA', 'WY']
print statesfilter
population[population.state.isin(statesfilter)][year]
population.query('state in @statesfilter')[['county_norm', year]]

population.county_norm.nunique()

indian_population[(indian_population.state=='WA') & (indian_population.county_norm=='king')]

us_counties[(us_counties.state=='WA') & (us_counties.county_norm=='king')]

donations[(donations.state=='WA') & (donations.county_norm=='king')].county_id.unique()
